A program that computes the first 50 fibonacci sequence

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	unsigned long long int num_1 = 1, num_2 = 2, sum, fibon;
+	unsigned long int num_1 = 1, num_2 = 2, sum, fibon;
 
 	printf("%lu, %lu", num_1, num_2);
 	for (fibon = 2; fibon < 50; fibon++)
